requirements.txt
flask==3.0.0
python-telegram-bot==20.7
newspaper3k==0.2.8
requests==2.31.0
python-dotenv==1.0.0
lxml==4.9.3
beautifulsoup4==4.12.2
nltk==3.8.1
aiohttp==3.9.1
APScheduler==3.10.4
